feat: implement mechanisms to handle KeyboardInterrupt
- Utilize a context manager to manage the Chrome webdriver.

- Incorporate a KeyboardInterrupt exception handler in
  'execute_scraper' to capture Ctrl+C exceptions for all
  store web scrapers.

- Register 'terminate_chromedriver_orphans' to run at the end
  of the process, ensuring the elimination of all remaining
  Chrome driver tasks.

+++ scraper/chrome_driver.py
@@ -4,6 +4,7 @@ import shutil
 import time
 
 import attr
+import psutil
 import requests
 from selenium import webdriver
 from selenium.common.exceptions import WebDriverException
@@ -12,9 +13,21 @@ from selenium.webdriver.chrome.service import Service as ChromeService
 from webdriver_manager.chrome import ChromeDriverManager
 
 
-@attr.s(slots=True, frozen=True, repr=False, eq=False, hash=False)
+class ChromeDriverError(Exception):
+    pass
+
+
+@attr.s(slots=True, frozen=False, repr=False, eq=False, hash=False)
 class ChromeDriver:
     cache_dir: str = attr.ib(default="chrome_cache")
+    driver: webdriver = attr.ib(default=None)
+
+    def __enter__(self):
+        self.initial()
+        return self
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        self.cleanup()
 
     def initial(self):
         ChromeDriverManager().install()
@@ -22,8 +35,21 @@ class ChromeDriver:
             os.mkdir(self.cache_dir)
 
     def cleanup(self):
+        self.terminate_chromedriver_orphans()
         if os.path.exists(self.cache_dir):
-            shutil.rmtree(self.cache_dir)
+            try:
+                shutil.rmtree(self.cache_dir)
+            except OSError as e:
+                if e.winerror == 32:
+                    print("Unable to delete the file or folder as it is being used by another process.")
+                    print("Terminate 'chromedriver.exe' orphans again.")
+                    self.terminate_chromedriver_orphans()
+                    shutil.rmtree(self.cache_dir)
+
+    def quit(self):
+        if self.driver:
+            self.driver.quit()
+            self.driver = None
 
     def create(self, max_retries=3, retry_delay_sec=3) -> webdriver:
         user_data_dir = os.path.join(self.cache_dir, f"pid_{os.getpid()}")
@@ -32,26 +58,34 @@ class ChromeDriver:
         chrome_options.add_argument("--headless")  # Headless mode, no browser window displayed
         chrome_options.add_argument("--disable-gpu")  # Disable GPU acceleration
         chrome_options.add_argument('--disable-logging')  # Disable JavaScript frontend logs
-
         chrome_options.add_argument(f"--user-data-dir={user_data_dir}")
         chrome_options.executable_path = os.path.join(user_data_dir, "chromedriver")
 
         for retry in range(max_retries):
             try:
-                # Set Chrome log level to silent
                 chrome_service = ChromeService()
                 chrome_service.silent = True  # Set silent to True to disable logging
-
-                # Create an instance of Chrome browser
                 driver = webdriver.Chrome(service=chrome_service, options=chrome_options)
-
-                # Wait for up to 5 seconds for the element to appear, throw an exception if not found.
-                # driver.implicitly_wait(5)
-
+                self.driver = driver
                 return driver
             except (WebDriverException, requests.exceptions.ConnectionError) as e:
                 print(f"Error: {e}. attempt {retry + 1}/{max_retries}, retrying...")
                 time.sleep(retry_delay_sec)
-
         else:
-            raise Exception(f"Failed to create Chrome WebDriver after {max_retries} attempts.")
+            raise ChromeDriverError(f"Failed to create Chrome WebDriver after {max_retries} attempts.")
+
+    @staticmethod
+    def terminate_chromedriver_orphans():
+        for proc in psutil.process_iter(attrs=['pid', 'name']):
+            try:
+                if proc.name() == 'chromedriver.exe':
+                    parent = psutil.Process(proc.pid)
+                    children = parent.children(recursive=True)
+                    for child in children:
+                        child.terminate()
+                        child.wait()
+                    parent.terminate()
+                    parent.wait()
+                    print(f"Terminate 'chromedriver.exe' process tree with PID {proc.pid}")
+            except (psutil.NoSuchProcess, psutil.AccessDenied, psutil.ZombieProcess):
+                pass
